Functions declared in RvSpectMLBase and specialized for each instrument

Functions declared in RvSpectMLBase and specialized for each instrument

Instrument-specific traits

Instrument-specific computation

Todo

Figure out what Documenter can't find the docstrings for these. They're in src/instruments/common.jl.

Instrument-specific traits

RvSpectMLBase.InstrumentsCommon.min_order — Function

Returns minimum allowable order for the specified instrument

RvSpectMLBase.InstrumentsCommon.max_order — Function

Returns maximum allowable order for the specified instrument

RvSpectMLBase.InstrumentsCommon.orders_all — Function

Returns range of all allowable orders for the specified instrument

RvSpectMLBase.InstrumentsCommon.orders_to_use_default — Function

Returns range of order indices to be used by default for the specified instrument

RvSpectMLBase.InstrumentsCommon.min_pixel_in_order — Function

Returns minimum allowable pixel index for the specified instrument and order index

RvSpectMLBase.InstrumentsCommon.max_pixel_in_order — Function

Returns maximum allowable pixel index for the specified instrument and order index

RvSpectMLBase.InstrumentsCommon.min_pixel — Function

Returns minimum allowable pixel index for the specified instrument and any order

RvSpectMLBase.InstrumentsCommon.max_pixel — Function

Returns maximum allowable pixel index for the specified instrument and any order

RvSpectMLBase.InstrumentsCommon.pixels_all — Function

Returns range of all allowable pixels for the specified instrument

RvSpectMLBase.InstrumentsCommon.min_pixels_in_chunk — Function

Returns maximum number of pixels in a chunk from the specified instrument

RvSpectMLBase.InstrumentsCommon.max_pixels_in_spectra — Function

Returns maximum number of pixels in a spectrum for the specified instrument

RvSpectMLBase.InstrumentsCommon.min_col_default — Function

Returns minimum pixel index to be used by default for the specified instrument

RvSpectMLBase.InstrumentsCommon.max_col_default — Function

Returns maximum pixel index to be read from header into metadata by default for the specified instrument

RvSpectMLBase.InstrumentsCommon.metadata_symbols_default — Function

Returns array of symbols to be stored from metadata in header by default for the specified instrument

RvSpectMLBase.InstrumentsCommon.metadata_strings_default — Function

Returns array of strings to be read from header into metadata by default for the specified instrument

RvSpectMLBase.InstrumentsCommon.default_ccf_mask_v_width — Function

Returns the default velocity width (in m/s) of a tophat CCF mask for the specified instrument

RvSpectMLBase.InstrumentsCommon.get_inst_module — Function

Returns the module containing code specific to the specified instrument

Instrument-specific computation

RvSpectMLBase.InstrumentsCommon.get_λ_range — Function

Returns the maximum range of wavelengths (in solar system barycenter frame) to be used for given spectrum

RvSpectMLBase.InstrumentsCommon.filter_line_list — Function

filter_line_list( df, inst ) Returns a dataframe filtered to only include lines appropriate for the specified instrument. Inputs:

df: Dataframe containing a column with key :lambda for wavelengths to be considered.

inst: instrument whose properties will be used in choosing which lines to keep.

Optional Inputs:

λmin: Overide the instrument's default minimum wavelength

λmax: Overide the instrument's default maximum wavelength