RvSpectMLBase Modules

RvSpectMLBase Modules

Instrument-specific Modules

Other Modules

Instrument-specific Modules

RvSpectMLBase.TheoreticalInstrument — Module

Module providing types and traits and customized functions for a generic theoretical spectrograph.

Other Modules

RvSpectMLBase.Pipeline — Module

The PipelinePlan stores what work needs to be done, what plots should be made, and what data/plots should be saved to disk. Users will querty it via needto(plan, symbol), makeplot(plan, symbol), saveplot(plan, symbol), and savedata(plan, symbol). Users can update the plan with makeplot!(plan, symbol), dontmake_plot!(plan, symbol), etc.

In principle, we could cache data here, too. But I'm not sure if we'll actulaly use that. So consider the cache as experimental.