Types Exported by RvSpectMLBase

Types Exported by RvSpectMLBase

Abstract Types

General purpose

Instrument specific

Other

Abstract Types

RvSpectMLBase.AbstractChunkOfSpectrum — Type

Abstract type for any ChunkOfSpectrum

RvSpectMLBase.AbstractInstrument — Type

Abstract Base type for Instruments

RvSpectMLBase.AbstractInstrument1D — Type

Abstract Base type for 1-D extracted spectra Should be specialized for specific instruments.

RvSpectMLBase.AbstractInstrument2D — Type

Abstract Base type for 2-D extracted spectra. Should be specialized for each instrument

RvSpectMLBase.AbstractSpectra — Type

Abstract type for any Spectrum (or region of spectrum)

RvSpectMLBase.AbstractSpectra1D — Type

Abstract type for any 1-d Spectrum (or region of spectrum)

RvSpectMLBase.AbstractSpectra2D — Type

Abstract type for any 2-d Spectrum

RvSpectMLBase.AbstractSpectralTimeSeriesCommonWavelengths — Type

Abstract type for a time series of spectra that share a common wavelength grid.

General purpose

RvSpectMLBase.ChunkList — Type

Struct containing an array of ChunkOfSpectrum

RvSpectMLBase.ChunkListTimeseries — Type

Mtching lists of times and array of ChunkLists

RvSpectMLBase.ChunkOfSpectrum — Type

ChunkOfSpectrum for views into Spectra1DBasic or Spectra2DBasic

RvSpectMLBase.Spectra1DBasic — Type

Basic struct for Spectra1D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

RvSpectMLBase.Spectra2DBasic — Type

Basic struct for Spectra2D (or region of specturm) Instruments can specialize their own if additional data is avaliable.

RvSpectMLBase.SpectralTimeSeriesCommonWavelengths — Type

Time series of spectra that share a common wavelength grid.

Instrument specific

RvSpectMLBase.TheoreticalInstrument.TheoreticalInstrument1D — Type

Trait for 1D spectra from a theoretical instrument

RvSpectMLBase.TheoreticalInstrument.TheoreticalInstrument2D — Type

Trait for a 2D spectra from a theoretical instrument

Other